+++ app/src/main/java/carnerero/agustin/cuentaappandroid/domain/database/accountusecase/UpdateAccountBalanceByExchangeRateUseCase.kt
@@ -1,11 +1,12 @@
 package carnerero.agustin.cuentaappandroid.domain.database.accountusecase
 
 import carnerero.agustin.cuentaappandroid.data.repository.AccountRepository
+import java.math.BigDecimal
 import javax.inject.Inject
 
 class UpdateAccountBalanceByExchangeRateUseCase  @Inject constructor(private val repository: AccountRepository){
 
-    suspend operator fun invoke(rate: Double) {
+    suspend operator fun invoke(rate: BigDecimal) {
         return repository.updateAccountBalanceByExchangeRate (rate)
     }
 }
