+++ app/src/main/java/carnerero/agustin/cuentaappandroid/domain/database/entriesusecase/GetSumTotalExpensesUseCase.kt
@@ -1,10 +1,11 @@
 package carnerero.agustin.cuentaappandroid.domain.database.entriesusecase
 
 import carnerero.agustin.cuentaappandroid.data.repository.EntryRepository
+import java.math.BigDecimal
 import javax.inject.Inject
 
 class GetSumTotalExpensesUseCase  @Inject constructor(private val repository: EntryRepository){
 
-    suspend operator fun invoke():Double = repository.getSumExpenses()
+    suspend operator fun invoke(): BigDecimal = repository.getSumExpenses()
 
 }
